ELEC-37 Reverted pqueue size.

+++ libraries/ms-common/inc/pqueue.h
@@ -21,7 +21,7 @@ typedef struct PQueue {
 void pqueue_init(PQueue *queue, PQueueNode *nodes, size_t num_nodes);
 
 // Push a node with the specified priority and data onto the queue.
-StatusCode pqueue_push(PQueue *queue, void *data, uint32_t prio);
+StatusCode pqueue_push(PQueue *queue, void *data, uint16_t prio);
 
 // Pop the minimum node from the queue and return its data.
 void *pqueue_pop(PQueue *queue);
