+++ libraries/ms-common/inc/pqueue_backed.h
@@ -20,7 +20,7 @@ StatusCode pqueue_backed_init_impl(PQueueBacked *queue, PQueueNode *nodes, void 
                                    size_t num_nodes, size_t num_elems, size_t elem_size);
 
 // Push a copy of the data in elem with the specified priority onto the pqueue.
-StatusCode pqueue_backed_push(PQueueBacked *queue, void *elem, uint32_t prio);
+StatusCode pqueue_backed_push(PQueueBacked *queue, void *elem, uint16_t prio);
 
 // Pop the minimum node from the pqueue and copy its data into elem.
 StatusCode pqueue_backed_pop(PQueueBacked *queue, void *elem);
